+++ ams/ams-server/src/test/java/com/netease/arctic/ams/server/AmsTestBase.java
@@ -56,7 +56,6 @@ import com.netease.arctic.ams.server.service.impl.FileInfoCacheService;
 import com.netease.arctic.ams.server.service.impl.JDBCMetaService;
 import com.netease.arctic.ams.server.service.impl.OptimizeQueueService;
 import com.netease.arctic.ams.server.service.impl.OptimizerService;
-import com.netease.arctic.ams.server.service.impl.OrphanFilesCleanServiceTest;
 import com.netease.arctic.ams.server.service.impl.PlatformFileInfoService;
 import com.netease.arctic.ams.server.util.DerbyTestUtil;
 import com.netease.arctic.ams.server.utils.CatalogUtil;
@@ -115,7 +114,6 @@ import static org.powermock.api.mockito.PowerMockito.when;
     TestIcebergMinorOptimizeCommit.class,
     TestExpireFileCleanSupportIceberg.class,
     TestOrphanFileCleanSupportIceberg.class,
-    OrphanFilesCleanServiceTest.class,
     TestOrphanFileClean.class,
     TestFileInfoCacheService.class,
     SupportHiveTestGroup.class,
