--- ams/ams-server/src/test/java/com/netease/arctic/ams/server/service/impl/OrphanFilesCleanServiceTest.java
@@ -1,35 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.netease.arctic.ams.server.service.impl;
-
-import com.netease.arctic.utils.FileUtil;
-import org.junit.Assert;
-import org.junit.Test;
-
-public class OrphanFilesCleanServiceTest {
-  @Test
-  public void testGetUriPath() {
-    Assert.assertEquals("/a/b/c", FileUtil.getUriPath("hdfs://xxxxx/a/b/c"));
-    Assert.assertEquals("/a/b/c", FileUtil.getUriPath("hdfs://localhost:8888/a/b/c"));
-    Assert.assertEquals("/a/b/c", FileUtil.getUriPath("file://xxxxx/a/b/c"));
-    Assert.assertEquals("/a/b/c", FileUtil.getUriPath("/a/b/c"));
-    Assert.assertEquals("/a/b/c", FileUtil.getUriPath("hdfs:/a/b/c"));
-    Assert.assertEquals("a/b/c", FileUtil.getUriPath("a/b/c"));
-  }
-}
